Breaking the unit testing classes into two files since file length was getting long
