+++ travis/unit-tests.sh
@@ -4,7 +4,7 @@ export PYTHONPATH=$PWD
 coverage run \
     --include='archiveinterface/*' \
     --omit='archiveinterface/archivebackends/abstract/*' \
-    -m pytest archiveinterface/archive_interface_unit_tests.py -v
+    -m pytest -p -v archiveinterface/archive_interface_unit_tests.py archiveinterface/archive_posix_unit_tests.py
 coverage report -m --fail-under 70
 if [[ $CODECLIMATE_REPO_TOKEN ]] ; then
   codeclimate-test-reporter
